+++ back/server/api/delete_url.ts
@@ -3,6 +3,9 @@ import { logError } from '../helpers/logger';
 
 export const delete_url = async (req: any, res: any): Promise<void> => {
   try {
+    if (req.session.isReadOnlySuperadmin) {
+      return res.status(401).end();
+    }
     const urlId = req.params.id;
     await db.query(`DELETE FROM url_list WHERE id=$1 AND group_id=$2`, [
       urlId,
@@ -10,7 +13,7 @@ export const delete_url = async (req: any, res: any): Promise<void> => {
     ]);
     res.status(200).end();
   } catch (e) {
-    logError("delete_url", e);
+    logError('delete_url', e);
     res.status(400).end();
   }
 };
